requirements.txt
argcomplete==1.10.0
attrs==19.3.0
beautifulsoup4==4.8.0
chardet==3.0.4
Click==7.0
doc2text==0.2.4
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
flasgger==0.9.3
Flask==1.1.1
future==0.18.2
IMAPClient==2.1.0
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
lxml==4.4.1
MarkupSafe==1.1.1
mime==0.1.0
mistune==0.8.4
more-itertools==7.2.0
nltk==3.4.5
numpy==1.17.4
olefile==0.46
pdfminer.six==20181108
Pillow==6.2.1
pycryptodome==3.9.4
PyPDF2==1.26.0
pyrsistent==0.15.6
pytesseract==0.3.0
python-pptx==0.6.18
python-rake==1.4.5
pytz==2019.3
PyYAML==5.1.2
rake-nltk==1.0.4
scipy==1.3.3
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
textract==1.6.3
tzlocal==1.5.1
Werkzeug==0.16.0
xlrd==1.2.0
XlsxWriter==1.2.6
zipp==0.6.0
